Implement threshold-based ECN marking
This commit implements DCTCP-style threshold-based ECN marking in
FullNoteLockQueue and MarkIPCE. When a packet arrives at a
FullNoteLockQueue, if the queue's size is greater than a threshold,
then a user annotation is set. MarkIPCE only sets a packet's ECN bits
to "Congestion Experienced" if that user annotation has been set.